+++ test/xcop.test.js
@@ -23,6 +23,7 @@ const PACKS = [
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'node-set-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'count-packs')),
   ...allFilesFrom(path.resolve(__dirname, 'resources', 'string-length-packs')),
+  ...allFilesFrom(path.resolve(__dirname, 'resources', 'name-packs')),
 ]
 
 /**